Test assistant with ace editor
