+++ package.json
@@ -10,7 +10,7 @@
     "babel-loader": "^8.1.0",
     "cross-env": "^5.2.0",
     "css-loader": "^2.1.1",
-    "jest": "^25.5.4",
+    "jest": "^26.5.3",
     "mini-css-extract-plugin": "^0.6.0",
     "n3": "^1.6.3",
     "raw-loader": "^4.0.1",
